# Improve categorical cast performance 