Benchmarks of casting `Utf8Chunked` with a size of 100k and random string lengths 2 < n < 80.

```
bench                   time:   [8.6611 ms 8.6646 ms 8.6683 ms]                  
                        change: [-77.388% -77.175% -76.962%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) low severe
  1 (1.00%) low mild
  5 (5.00%) high mild
  2 (2.00%) high severe

bench_global_cache      time:   [11.896 ms 11.937 ms 11.980 ms]                               
                        change: [-64.854% -64.429% -64.020%] (p = 0.00 < 0.05)
                        Performance has improved.


```